+++ NEWS.md
@@ -111,6 +111,7 @@ New library features
 * `Lockable` is now exported ([#54595])
 * `Base.require_one_based_indexing` and `Base.has_offset_axes` are now public ([#56196])
 * New `ltruncate`, `rtruncate` and `ctruncate` functions for truncating strings to text width, accounting for char widths ([#55351])
+* `Threads.@spawn` can now take a `reserved_stack` argument to set a stack size limit ([#55201])
 * `isless` (and thus `cmp`, sorting, etc.) is now supported for zero-dimensional `AbstractArray`s ([#55772])
 * `invoke` now supports passing a Method instead of a type signature making this interface somewhat more flexible for certain uncommon use cases ([#56692]).
 * `Timer(f, ...)` will now match the stickiness of the parent task when creating timer tasks, which can be overridden